mcp-script/requirements.txt
# Host-side dependencies for the local MCP <-> llama.cpp bridge.
# Note: This is separate from rag-service/requirements.txt.

mcp
openai
psycopg[binary]
python-dotenv
sqlparse
anyio
fastapi
uvicorn
